pyproject.toml
[project]
name = "base"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.32.0",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
]
